pyproject.toml
[project]
name = "pystack"
dynamic = ["version"]
description = "A unified educational platform integrating PyOS, Pebble, and PyDB"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    { name = "Rob Cox" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Education",
    "Typing :: Typed",
]
keywords = ["education", "full-stack", "operating-system", "compiler", "database"]
dependencies = [
    "pyos-learn",
    "pebble-lang",
    "pydb",
]

[project.urls]
Homepage = "https://github.com/robcox2910/pystack"
Repository = "https://github.com/robcox2910/pystack"

[project.scripts]
pystack = "pystack.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.3",
    "pre-commit>=4.0.1",
    "pyright>=1.1.400",
    "commitizen>=4.4.0",
    "pip-audit>=2.10.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/pystack/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/pystack"]

[tool.uv.sources]
pyos-learn = { path = "../os_learning", editable = true }
pebble-lang = { path = "../pebble-lang", editable = true }
pydb = { path = "../pydb", editable = true }

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py314"
src = ["src", "tests"]
exclude = ["src/pystack/_version.py"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "UP", "B", "C4", "C90", "PTH", "SIM",
    "TID", "S", "PERF", "FURB", "PIE", "T10", "T20", "ERA", "PL",
    "RET", "ARG", "FBT", "DTZ", "ICN", "PT", "D", "RUF", "A", "BLE",
    "EM", "FLY", "ISC", "PGH", "RSE", "SLF", "TRY",
]
ignore = [
    "E501", "D203", "D213", "ISC001", "TRY003",
    "PIE807", "PLW0108", "FBT001", "FBT003",
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pylint]
max-args = 9

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "SLF001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src/pystack",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*", "*/_version.py"]

[tool.coverage.report]
precision = 2
show_missing = true

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
include = ["src", "tests"]
venvPath = "."
venv = ".venv"

[[tool.pyright.executionEnvironments]]
root = "tests"
extraPaths = ["."]
reportPrivateUsage = false

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "scm"
